Progress (Live Editing)

+++ src/adt/jam/jam.h
@@ -63,7 +63,7 @@ void TulisJam (JAM J);
 /* ***************************************************************** */
 /* KELOMPOK KONVERSI TERHADAP TYPE                                   */
 /* ***************************************************************** */
-long JamToDetik (JAM J);
+int JamToDetik (JAM J);
 /* Diberikan sebuah JAM, mengkonversi menjadi jumlah detik dari pukul 0:0:0 */
 /* Rumus : detik = 3600*HH + 60*MM + SS */
 /* Nilai maksimum = 3600*23+59*60+59 */
